backend/scraper/resume_parser.py
"""
Resume Parser for AutoPortfolio
Extracts information from PDF and DOCX resume files
"""

import os
import re
from pathlib import Path
import logging

# PDF parsing
try:
    import PyPDF2
    PDF_AVAILABLE = True
except ImportError:
    PDF_AVAILABLE = False

# DOCX parsing
try:
    from docx import Document
    DOCX_AVAILABLE = True
except ImportError:
    DOCX_AVAILABLE = False

logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)


class ResumeParser:
    """Parse resume files to extract structured information"""
    
    def __init__(self):
        self.skills_keywords = [
            # Programming Languages
            'python', 'javascript', 'java', 'c++', 'c#', 'php', 'ruby', 'go', 'rust', 'swift',
            'kotlin', 'typescript', 'scala', 'r', 'matlab', 'sql', 'html', 'css',
            
            # Frameworks & Libraries
            'react', 'angular', 'vue', 'node.js', 'express', 'django', 'flask', 'spring',
            'laravel', 'rails', 'asp.net', 'jquery', 'bootstrap', 'tailwind',
            
            # Databases
            'mysql', 'postgresql', 'mongodb', 'redis', 'sqlite', 'oracle', 'cassandra',
            'elasticsearch', 'dynamodb',
            
            # Cloud & DevOps
            'aws', 'azure', 'gcp', 'docker', 'kubernetes', 'jenkins', 'git', 'github',
            'gitlab', 'bitbucket', 'terraform', 'ansible', 'nginx', 'apache',
            
            # Tools & Technologies
            'linux', 'windows', 'macos', 'bash', 'powershell', 'vim', 'vscode',
            'intellij', 'eclipse', 'postman', 'jira', 'confluence', 'slack',
            
            # Data & AI
            'pandas', 'numpy', 'scikit-learn', 'tensorflow', 'pytorch', 'keras',
            'opencv', 'matplotlib', 'seaborn', 'jupyter', 'tableau', 'powerbi'
        ]
    
    def parse_resume(self, file_path):
        """Parse resume file and extract structured information"""
        try:
            file_path = Path(file_path)
            
            if not file_path.exists():
                logger.error(f"File not found: {file_path}")
                return self._get_dummy_resume_data()
            
            # Extract text based on file type
            if file_path.suffix.lower() == '.pdf':
                text = self._extract_pdf_text(file_path)
            elif file_path.suffix.lower() in ['.docx', '.doc']:
                text = self._extract_docx_text(file_path)
            else:
                logger.warning(f"Unsupported file type: {file_path.suffix}")
                return self._get_dummy_resume_data()
            
            if not text:
                logger.warning("No text extracted from resume")
                return self._get_dummy_resume_data()
            
            # Parse extracted text
            parsed_data = self._parse_text(text)
            logger.info("✅ Resume parsed successfully")
            return parsed_data
            
        except Exception as e:
            logger.error(f"Error parsing resume: {str(e)}")
            return self._get_dummy_resume_data()
    
    def _extract_pdf_text(self, file_path):
        """Extract text from PDF file"""
        if not PDF_AVAILABLE:
            logger.warning("PyPDF2 not available, cannot parse PDF")
            return ""
        
        try:
            text = ""
            with open(file_path, 'rb') as file:
                pdf_reader = PyPDF2.PdfReader(file)
                for page in pdf_reader.pages:
                    text += page.extract_text() + "\n"
            return text
        except Exception as e:
            logger.error(f"Error extracting PDF text: {str(e)}")
            return ""
    
    def _extract_docx_text(self, file_path):
        """Extract text from DOCX file"""
        if not DOCX_AVAILABLE:
            logger.warning("python-docx not available, cannot parse DOCX")
            return ""
        
        try:
            doc = Document(file_path)
            text = ""
            for paragraph in doc.paragraphs:
                text += paragraph.text + "\n"
            return text
        except Exception as e:
            logger.error(f"Error extracting DOCX text: {str(e)}")
            return ""
    
    def _parse_text(self, text):
        """Parse extracted text to find structured information"""
        text_lower = text.lower()
        
        return {
            'skills': self._extract_skills(text_lower),
            'experience': self._extract_experience(text),
            'education': self._extract_education(text),
            'contact': self._extract_contact_info(text)
        }
    
    def _extract_skills(self, text_lower):
        """Extract skills from text"""
        found_skills = []
        
        for skill in self.skills_keywords:
            # Look for whole word matches
            pattern = r'\b' + re.escape(skill.lower()) + r'\b'
            if re.search(pattern, text_lower):
                found_skills.append(skill.title())
        
        # Remove duplicates and return
        return list(set(found_skills))
    
    def _extract_experience(self, text):
        """Extract work experience from text"""
        experience = []
        
        # Look for common experience patterns
        experience_patterns = [
            r'(\d{4})\s*[-–]\s*(\d{4}|\w+)\s*[:\-]?\s*([^\n]+)',
            r'(\w+\s+\d{4})\s*[-–]\s*(\w+\s+\d{4}|\w+)\s*[:\-]?\s*([^\n]+)',
        ]
        
        for pattern in experience_patterns:
            matches = re.finditer(pattern, text, re.MULTILINE | re.IGNORECASE)
            for match in matches:
                start_date = match.group(1)
                end_date = match.group(2)
                description = match.group(3).strip()
                
                if len(description) > 10:  # Filter out short matches
                    experience.append({
                        'duration': f"{start_date} - {end_date}",
                        'description': description[:200]  # Limit length
                    })
        
        return experience[:5]  # Return top 5 experiences
    
    def _extract_education(self, text):
        """Extract education information from text"""
        education = []
        
        # Common education keywords
        education_keywords = [
            'university', 'college', 'institute', 'school', 'bachelor', 'master',
            'phd', 'degree', 'diploma', 'certification', 'course'
        ]
        
        lines = text.split('\n')
        for line in lines:
            line_lower = line.lower()
            if any(keyword in line_lower for keyword in education_keywords):
                if len(line.strip()) > 10:  # Filter out short lines
                    education.append(line.strip()[:100])  # Limit length
        
        return education[:3]  # Return top 3 education entries
    
    def _extract_contact_info(self, text):
        """Extract contact information from text"""
        contact = {}
        
        # Email pattern
        email_pattern = r'\b[A-Za-z0-9._%+-]+@[A-Za-z0-9.-]+\.[A-Z|a-z]{2,}\b'
        email_match = re.search(email_pattern, text)
        if email_match:
            contact['email'] = email_match.group()
        
        # Phone pattern (basic)
        phone_pattern = r'[\+]?[1-9]?[0-9]{7,15}'
        phone_match = re.search(phone_pattern, text)
        if phone_match:
            contact['phone'] = phone_match.group()
        
        # LinkedIn URL pattern
        linkedin_pattern = r'linkedin\.com/in/[\w\-]+'
        linkedin_match = re.search(linkedin_pattern, text, re.IGNORECASE)
        if linkedin_match:
            contact['linkedin'] = f"https://{linkedin_match.group()}"
        
        return contact
    
    def _get_dummy_resume_data(self):
        """Return dummy resume data for demo purposes"""
        return {
            'skills': [
                'Python', 'JavaScript', 'React', 'Node.js', 'Django', 
                'PostgreSQL', 'AWS', 'Docker', 'Git', 'Machine Learning'
            ],
            'experience': [
                {
                    'duration': '2022 - Present',
                    'description': 'Senior Software Engineer at TechCorp - Led development of scalable web applications'
                },
                {
                    'duration': '2020 - 2022',
                    'description': 'Full Stack Developer at StartupXYZ - Built responsive web applications using modern frameworks'
                }
            ],
            'education': [
                'Bachelor of Science in Computer Science, University of Technology',
                'Full Stack Web Development Bootcamp, CodeAcademy'
            ],
            'contact': {
                'email': 'john.doe@example.com',
                'phone': '+1-555-0123',
                'linkedin': 'https://linkedin.com/in/johndoe'
            }
        }