backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
playwright
requests==2.31.0
python-dotenv==1.0.0
APScheduler==3.10.4
PyPDF2==3.0.1
python-docx==1.1.0
Jinja2==3.1.2